pyproject.toml
[tool.poetry]
name = "bottelechaat"
version = "0.1.0"
description = "Telegram bot for forwarding OpenAI verification codes"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "2.107.0"
google-auth-httplib2 = "0.1.1"
google-auth-oauthlib = "1.1.0"
python-telegram-bot = "13.15"
python-dotenv = "1.0.0"
flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 